refactor: Split AstraMisc.h into .h/.cpp
